move expect stream trait
