+++ utils-aio/src/duplex.rs
@@ -2,61 +2,13 @@ use std::{
     io,
     io::{Error, ErrorKind},
     pin::Pin,
-    task::{Context, Poll},
 };
 
-use futures::{
-    channel::mpsc, future::FusedFuture, stream::FusedStream, AsyncRead, AsyncWrite, Future, Sink,
-    Stream, TryStream, TryStreamExt,
-};
+use futures::{channel::mpsc, AsyncRead, AsyncWrite, Sink, Stream};
 
 pub trait DuplexByteStream: AsyncWrite + AsyncRead + Unpin {}
 
 impl<T> DuplexByteStream for T where T: AsyncWrite + AsyncRead + Unpin {}
-
-/// Future for the [`expect_next`](Duplex::expect_next) method.
-#[derive(Debug)]
-#[must_use = "futures do nothing unless you `.await` or poll them"]
-pub struct ExpectNext<'a, St: ?Sized> {
-    stream: &'a mut St,
-}
-
-impl<St: ?Sized + Unpin> Unpin for ExpectNext<'_, St> {}
-
-impl<'a, St: ?Sized + TryStream + Unpin> ExpectNext<'a, St> {
-    pub(super) fn new(stream: &'a mut St) -> Self {
-        Self { stream }
-    }
-}
-
-impl<St: ?Sized + TryStream + Unpin + FusedStream> FusedFuture for ExpectNext<'_, St>
-where
-    <St as TryStream>::Error: Into<io::Error>,
-{
-    fn is_terminated(&self) -> bool {
-        self.stream.is_terminated()
-    }
-}
-
-impl<St: ?Sized + TryStream + Unpin> Future for ExpectNext<'_, St>
-where
-    <St as TryStream>::Error: Into<io::Error>,
-{
-    type Output = io::Result<St::Ok>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        self.stream
-            .try_poll_next_unpin(cx)
-            .map_err(|e| e.into())?
-            .map(|item| {
-                if let Some(item) = item {
-                    Ok(item)
-                } else {
-                    Err(io::ErrorKind::UnexpectedEof.into())
-                }
-            })
-    }
-}
 
 /// A channel that can be used to send and receive messages.
 pub trait Duplex<T>:
@@ -66,39 +18,6 @@ pub trait Duplex<T>:
     + Sync
     + Unpin
 {
-    /// Creates a future that attempts to resolve the next item in the stream.
-    /// If an error is encountered before the next item, the error is returned
-    /// instead.
-    ///
-    /// Additionally, if the stream ends before the next item, an error is
-    /// returned.
-    ///
-    /// This is similar to the [`TryStreamExt::try_next`](futures::stream::TryStreamExt::try_next)
-    /// combinator, but returns an error if the stream ends before the next item instead of an
-    /// `Option`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # futures::executor::block_on(async {
-    /// use futures::{SinkExt, StreamExt};
-    /// use utils_aio::duplex::{Duplex, MpscDuplex};
-    ///
-    /// let (mut a, mut b) = MpscDuplex::new();
-    ///
-    /// a.send(()).await.unwrap();
-    /// a.close().await.unwrap();
-    ///
-    /// assert!(b.expect_next().await.is_ok());
-    /// assert!(b.expect_next().await.is_err());
-    /// # })
-    /// ```
-    fn expect_next(&mut self) -> ExpectNext<'_, Self>
-    where
-        Self: Sized,
-    {
-        ExpectNext::new(self)
-    }
 }
 
 impl<T, U> Duplex<T> for U where
