+++ utils-aio/src/lib.rs
@@ -9,3 +9,4 @@ pub mod factory;
 #[cfg(feature = "mux")]
 pub mod mux;
 pub mod non_blocking_backend;
+pub mod stream;
